refactor: bookmark type - enum class
bookmark type - string to enum class
