+++ BeginVegan/presentation/src/main/java/com/beginvegan/presentation/config/bookmark/BookmarkModule.kt
@@ -1,6 +1,7 @@
 package com.beginvegan.presentation.config.bookmark
 
 import com.beginvegan.domain.useCase.bookmarks.BookmarkUseCase
+import com.beginvegan.presentation.config.enumclass.Bookmarks
 import com.beginvegan.presentation.util.BookmarkController
 import dagger.Module
 import dagger.Provides
@@ -13,29 +14,30 @@ import kotlinx.coroutines.coroutineScope
 object BookmarkModule {
 
     @Provides
-    fun provideBookmarkController(bookmarkUseCase:BookmarkUseCase):BookmarkController = object:BookmarkController{
-        override suspend fun postBookmark(contentId:Int, contentType:String): Boolean {
-            var onSuccess = false
-            coroutineScope {
-                bookmarkUseCase.postBookmark(contentId, contentType).onSuccess {
-                    onSuccess = true
-                }.onFailure {
-                    onSuccess = false
+    fun provideBookmarkController(bookmarkUseCase: BookmarkUseCase): BookmarkController =
+        object : BookmarkController {
+            override suspend fun postBookmark(contentId: Int, contentType: Bookmarks): Boolean {
+                var onSuccess = false
+                coroutineScope {
+                    bookmarkUseCase.postBookmark(contentId, contentType.type).onSuccess {
+                        onSuccess = true
+                    }.onFailure {
+                        onSuccess = false
+                    }
                 }
+                return onSuccess
             }
-            return onSuccess
-        }
 
-        override suspend fun deleteBookmark(contentId:Int, contentType:String): Boolean {
-            var onSuccess = false
-            coroutineScope {
-                bookmarkUseCase.deleteBookmark(contentId, contentType).onSuccess {
-                    onSuccess = true
-                }.onFailure {
-                    onSuccess = false
+            override suspend fun deleteBookmark(contentId: Int, contentType: Bookmarks): Boolean {
+                var onSuccess = false
+                coroutineScope {
+                    bookmarkUseCase.deleteBookmark(contentId, contentType.type).onSuccess {
+                        onSuccess = true
+                    }.onFailure {
+                        onSuccess = false
+                    }
                 }
+                return onSuccess
             }
-            return onSuccess
         }
-    }
 }
