+++ BeginVegan/presentation/src/main/java/com/beginvegan/presentation/config/enumclass/Bookmarks.kt
@@ -1,0 +1,7 @@
+package com.beginvegan.presentation.config.enumclass
+
+enum class Bookmarks(val type:String) {
+    MAGAZINE("MAGAZINE"),
+    RECIPE("RECIPE"),
+    RESTAURANT("RESTAURANT")
+}
