+++ BeginVegan/presentation/src/main/java/com/beginvegan/presentation/config/enumclass/MainPages.kt
@@ -1,4 +1,4 @@
-package com.beginvegan.presentation.util
+package com.beginvegan.presentation.config.enumclass
 
 enum class MainPages {
     HOME,
