+++ BeginVegan/presentation/src/main/java/com/beginvegan/presentation/util/BookmarkController.kt
@@ -1,6 +1,8 @@
 package com.beginvegan.presentation.util
 
+import com.beginvegan.presentation.config.enumclass.Bookmarks
+
 interface BookmarkController {
-    suspend fun postBookmark(contentId:Int, contentType:String) : Boolean
-    suspend fun deleteBookmark(contentId:Int, contentType:String): Boolean
+    suspend fun postBookmark(contentId:Int, contentType:Bookmarks) : Boolean
+    suspend fun deleteBookmark(contentId:Int, contentType:Bookmarks): Boolean
 }
